Forum now works,
The forum topics are now functional and the topics only contains posts related to the topic.

I am not really using any primary/foreigng key regarding the IDs for the topics and posts as of yet since i did not get it to work.

This will be fixed later on.